deprecate TerminalWidgetImpl class
- the abstract class `TerminalWidget` is removed, and `TerminalWidgetImpl` renamed into `TerminalWidget`

- resolves #7490
